+++ src/neptune/new/internal/operation_processors/async_operation_processor.py
@@ -43,7 +43,6 @@ from neptune.new.internal.operation_processors.operation_processor import Operat
 from neptune.new.internal.operation_processors.operation_storage import OperationStorage
 from neptune.new.internal.threading.daemon import Daemon
 from neptune.new.internal.utils.logger import logger
-from neptune.new.sync.utils import create_dir_name
 
 _logger = logging.getLogger(__name__)
 
@@ -91,7 +90,7 @@ class AsyncOperationProcessor(OperationProcessor):
     @staticmethod
     def _init_data_path(container_id: UniqueId, container_type: ContainerType):
         now = datetime.now()
-        container_dir = f"{NEPTUNE_DATA_DIRECTORY}/{ASYNC_DIRECTORY}/{create_dir_name(container_type, container_id)}"
+        container_dir = f"{NEPTUNE_DATA_DIRECTORY}/{ASYNC_DIRECTORY}/{container_type.create_dir_name(container_id)}"
         data_path = f"{container_dir}/exec-{now.timestamp()}-{now.strftime('%Y-%m-%d_%H.%M.%S.%f')}"
         data_path = data_path.replace(" ", "_").replace(":", ".")
         return data_path