+++ src/neptune/new/internal/operation_processors/offline_operation_processor.py
@@ -28,7 +28,6 @@ from neptune.new.internal.id_formats import UniqueId
 from neptune.new.internal.operation import Operation
 from neptune.new.internal.operation_processors.operation_processor import OperationProcessor
 from neptune.new.internal.operation_processors.operation_storage import OperationStorage
-from neptune.new.sync.utils import create_dir_name
 
 
 class OfflineOperationProcessor(OperationProcessor):
@@ -44,7 +43,7 @@ class OfflineOperationProcessor(OperationProcessor):
 
     @staticmethod
     def _init_data_path(container_id: UniqueId, container_type: ContainerType):
-        return f"{NEPTUNE_DATA_DIRECTORY}/{OFFLINE_DIRECTORY}/{create_dir_name(container_type, container_id)}"
+        return f"{NEPTUNE_DATA_DIRECTORY}/{OFFLINE_DIRECTORY}/{container_type.create_dir_name(container_id)}"
 
     def enqueue_operation(self, op: Operation, wait: bool) -> None:
         self._queue.put(op)