+++ tests/neptune/new/cli/test_sync.py
@@ -18,11 +18,11 @@ from unittest.mock import MagicMock
 
 import pytest
 
+from neptune.new.cli.sync import SyncRunner
+from neptune.new.cli.utils import get_qualified_name
 from neptune.new.internal.container_type import ContainerType
 from neptune.new.internal.operation import Operation
-from neptune.new.sync import SyncRunner
-from neptune.new.sync.utils import get_qualified_name
-from tests.neptune.new.sync.utils import (
+from tests.neptune.new.cli.utils import (
     execute_operations,
     generate_get_metadata_container,
     prepare_deprecated_run,