+++ tests/neptune/new/cli/utils.py
@@ -18,6 +18,7 @@ import threading
 from pathlib import Path
 from typing import Optional
 
+from neptune.new.cli.utils import get_qualified_name
 from neptune.new.constants import (
     ASYNC_DIRECTORY,
     OFFLINE_DIRECTORY,
@@ -27,7 +28,6 @@ from neptune.new.internal.backends.api_model import ApiExperiment
 from neptune.new.internal.container_type import ContainerType
 from neptune.new.internal.disk_queue import DiskQueue
 from neptune.new.internal.utils.sync_offset_file import SyncOffsetFile
-from neptune.new.sync.utils import get_qualified_name
 from tests.neptune.new.utils.api_experiments_factory import (
     api_metadata_container,
     api_run,